Add a scale factor option
Adds the `--scale-factor` option.

The `--scale-factor` option allows users to specify a specific device scale factor as a float.

This is useful for testing specific high-definition displays. For example, the Google Pixel 6 has a pixel ratio of 2.625. Some modern phones have pixel ratios of 4.

The existing `--retina` option is a specific case of the new `--scale-factor` option.

The commit also includes a validation to prevent conflicts when using the `--scale-factor` and `--retina` options together.