Fix consistently failing test: Selector with a wait
On moderately slow networks, the "selector with a wait" test fails.

There is already a --wait-for test, so replacing the --wait test with --wait-for was a bad idea.

This change reuses the --wait-for test's HTML with a delayed div instead.

This way --wait with --selector is still tested.

Fixes simonw/shot-scraper#134